writer-identification/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
albumentations>=1.3.0
opencv-python>=4.6.0
pandas>=1.4.0 